+++ source/scheduler/SoftwareTimer.cpp
@@ -2,7 +2,7 @@
  * \file
  * \brief SoftwareTimer class implementation
  *
- * \author Copyright (C) 2015 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ * \author Copyright (C) 2015-2016 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
  *
  * \par License
  * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
@@ -23,10 +23,10 @@ SoftwareTimer::~SoftwareTimer()
 
 }
 
-int SoftwareTimer::start(const TickClock::duration duration)
+int SoftwareTimer::start(const TickClock::duration duration, const TickClock::duration period)
 {
 	const auto now = TickClock::now();
-	return start(now + duration + decltype(duration){1});
+	return start(now + duration + decltype(duration){1}, period);
 }
 
 /*---------------------------------------------------------------------------------------------------------------------+